packages/dagster_pyiceberg_polars/pyproject.toml
[project]
name = "dagster-pyiceberg-polars"
version = "0.0.0"
description = "Package for storing Polars DataFrames in PyIceberg tables."
authors = [
  {name="Jasper Ginn", email="jasperginn@gmail.com"}
]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "dagster-pyiceberg",
    "polars>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
